cephadm: move low level networking funcs to net_utils.py
Pair-programmed-with: Adam King <adking@redhat.com>
